Add option to export errors every Night
Add an option to the CMakeLists.txt file which gives the user the option to
create a delimited text file containing the errors that are displayed to the
dashboard.  This is done by repurposing the ReportXINDEX function and the
package globbing to write to a file rather than the terminal screen.  The
old list should be removed prior to the start to create a clean list.

The WriteErrors file is configured in the CMakeLists file to contain the
correct paths.  The actual writing is done when running cmake on the
configured file by a test.

Change-Id: Iee4f43e28ddc94b2c89a06a64a55848aa95d9556