Add test cases for UseCases
Move and rename the AddAppointment.py.in to a separate folder for UseCase
testing. Do the same move for the PostSetup.py.in

Create a CMakeLists.txt for the UseCases folder, which configures the two
files and sets up two simple MakeAppointment tests.
One test should be for a normal day, the other should try to schedule an
appointment for July 4,2012.

Add an option to the CMakeLists.txt called VISTA_USE_CASE_TESTING, to run
the UseCase testing.
Change the main CMakeLists to add the UseCases subdirectory if the
VISTA_USE_CASE_TESTING option is selected.