Update the ImportRG
Add a copy of the ZTLOAD1.m routine that has been correctly changed for
running the initialization of VistA on GT.M to a folder called Patches in
the main directory.

Modify the ImportRG.cmake.in to copy the ZTLOAD1.m to the VistA-FOIA
directory before the packing of the RO.  This should only happen on GT.M.

Add another execute process to Run the PostSetup.py script after the
running of the Initialization.  This creates the necessary resource for
the UseCase testing.

Comment out the add_test for the FT_CPRSPostImport as it is not needed.