Update patient addition files
Add an appointment answer for veterans, as the elegibility code changes
between veterans and non-veterans

Move the sign-in of the system manager to the addpatients file to be run
after adding each patient unless it is the last one.  This will let these
exit to the prompt and have a more useful exit.