Integrate VerifyOneAppointment with Scenario001
Add a configure_file command in UseCases/CMakeLists.txt to move the
VerifyOneAppointment to be with the other python files.

Add an extra return on the end of MakeAppointment to return to the prompt,
where the VerifyOneAppointment expects to be.

Add the command to run the VerifyOneAppointment script from within
Scenario001 after the MakeAppointment call, but before the rollback.