Update Scenario and AddClinics file
Now that more doctors are added, one primary isn't enough for all clinics.

Add lists to the beginning of AddClinics that contains the list of
Doctors names.  They should be differentiated by the passed 'state'
parameter.  Then use this list to pass the correct names to the
AddOneClinic function.  Remove the 'primary' argument that is passed to
the function.

Remove the 'SMITH,JAMES' from the call to the AddClinics in Scenario001.