Prepopulate the other scenario files
Add a file for the rest of the Scenarios (2-11) that contain the
instructions for adding the commands to the file and running the test.
Each file shoud now fail the a CTest with the same intentional failure
message.

Add the add_test commands for each file in the UseCases/CMakeLists.txt