Remove transaction processing from Scenario
Remove the TSTART and TROLLBACK commands from Scenario001.

Replace the TROLLBACK command with a call to a new python file, which
stops the instance of Cache, copies in a user-created backup database,
then restarts the instance of Cache.  This path is captured by setting
VISTA_BACKUP to be the path to the folder containing the cache.dat.