Add Coverage to the UnitTest and UseCases.
Add an option (USE_COVERAGE) for computing code coverage in the
CMakeLists.txt.

If the option is selected, print a warning message saying that the tests
may take longer and use more memory.  Also warn that parsing the coverage
will require a newer version of CMake (v2.8.9)

Add the commands to start/stop the two utilities that are used for
coverage in the OSEHRAHelper.py.  See the README for more detailed
information.

Add the startCoverage and stopCoverage commands to the UnitTest.py.in and
the Scenario001.py.in file. Each command should be under an if statment
checking the USE_COVERAGE value.