Add Cache Backup script and test
Add a script to the UseCase testing framework which will shut down the
instance and copy the CACHE.DAT to a user specified location.  It will
check the SHA1 sum of the original and the copy to ensure that the copy
was correct.

If it was: a gzip will be taken of the copy and stored with a timestamp.

If not, the Cache instance will remain down and a message will be printed.

Add a test that will run this script before the Scenario tests and make
all of the Scenario tests depend on it.

Change-Id: I2d9cf12910769baa081bfe1111a90005eecc15e6