Add TestHelper and update RAS tests
Add code to the TestHelper.py function that adds the TestSuiteDriver
functionality.  It is used to simplify the generation of RAS tests
such that a tool like the RASR can be utilized.

Update the current RAS tests to use this new TestSuiteDriver class

Add paramikoe and RemoteConnection python files, which are used to
have a secure connection to a remote VistA instance.

Change-Id: Ibe3bfa931f8444e758c111f8cd291907db942737