Add Scripts for OTJ Evidence creation
Add two python scripts to the /Scripts directory which are used to create
MCode evidence for OTJ peer and final reviews.

Run "python OTJKIDSUtilities.py -h" for usage and arguments

It will parse a given KIDS file while saving the routines into a given output
dir, load the KIDS distribution, print out information about the transport
global and install it.  Then, it will run XINDEX on each install within
the file, write out the %RCHECK and %RFIND for the entire system, then
parse those results to write out findings for the routines sent in the
KIDS patch.  It will also parse the XINDEX results for the variables that
are not explicitly killed.

Modify the DefaultKIDSBuildInstaller to take an additional argument for
writing out the information about the transport global. Write the
connection functions to save each part: checksums, print of transport
global, and  compare to system.

If the scripts are run in a GT.M environment,instead of running the
%RCHECK and %RFIND, print a message saying that those are Cache only
utilities.  The GT.M instances will run a REGEX matching function in
place of the RFind utility.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-80
Change-Id: I30a84712763536e95f9e16e997fe41f4af230c67