Integrate testing for SSEP
Integrate the testing that was created for the SSEP into the testing
harness.

Add the ZZUTM2MC.m test to a new testing folder in the RPC Broker
package.  This addition requires a change to the UnitTest.py.in file as
it requires variables not set up by the call to "Q^DI" but are set up by
the XUP menu.

Also restructure the PostImportSetupScript to ensure that a restart of
TaskMan is performed after setting the Listener to queue on startup.

Add the TestSSEP.py and necessary XML files to the PyUnit folder under
the Scripts/Testing directory.

Add checks before each new test to see if the RPC Listener has been
started.  If it hasn't, do not perform the test. The MUnit check can be
added to the .m file, while the SSEP should use a separate python function
to connect and run the same test.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-102
Change-Id: Iddf55c23daf65ae10d8807f6c701879dd77e6b42