Update ReportXINDEXErrors parsing logic
Add a '%' to the regular expressions that are used to capture the name of
the reported routine.  The parser would normally miss routines that
started with %, "%ZTLOAD" in the Kernel Package.

Also, add logic to the if statment that checks that an exemption was not
found to also quit if the routine name starts with "%Z....".  Since we
have exceptions for the prior-to-renaming routine, this is mostly
re-reporting errors.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-113
Change-Id: I065eefc7cdfd400c91276321034f5b0b9376eda5