Fix username/password bug for RAS tests
The code used in the OSEHRA RAS functional tests are unable to
run on an instance that uses a username/password prompt before displaying
the MUMPS prompt.

Change the RASTest.cmake.in to export the VISTA_CACHE_USERNAME and
VISTA_CACHE_PASSWORD value to environment variables.

Change the TestHelper code that connects to the instance to attempt to grab
the VISTA_CACHE_USERNAME value.  After connection, the test_suite_details.username
will be checked for a value.  If it exists, the code should grab the VISTA_CACHE_PASSWORD
value and use them to allow access to a MUMPS prompt by answering the signon prompts

OSEHRA-Id: