Add variable to limit Coverage Calculation.
Add a TEST_VISTA_COVERAGE_SUBSET_DIR advanced variable.  It should be a file path
to a directory that contains routine files (*.m).  If routines are found,
the names of the routines in that directory should be used as the routines
for the line-by-line monitor to pay attention to.

It should also be used as the "Packages" argument to the mcov/cmcov files for CTests
calculation of coverage.

The file names should also be passed to the Coverage Preparation test
which will only write out the files that are found in that directory.

Change the files that have a call to startCoverage to accept this
variable.  This is more complicated for the RAS tests, as it needs to be
added to the test_suite_details information and passed as a command line
argument

OSEHRA-Id: http://issues.osehra.org/browse/OAT-136
Change-Id: Ib432782fc37171a0656ae4adbd3d097d5ced5eef