Add OSEHRA changes to MUnit Test routines
Alter the MUnit test routines that are called to not run if %ut routine
does not exist.

Alter the %utt1 routine to not have %utt6 as an included routine through
XTROU.  This eliminates the two errors that are thrown when running the
%utt6 routines not "from the top"

Alter the ZZUTPOST routine to only rename the utt routines.
Also, eliminate the running of the %utt1 routine after the renaming.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-152
Change-Id: I9151370d03c281f3a912391b45ecd47d4904383e