Expand setup of fictional hospital users
Make changes to the set up of the users (Doctor, Nurse, Clerk) for the
fictional hospital set up in the PostImportSetupScript.

  Add the nurse user to the NURS STAFF file via the "Staff Record Edit"
  option

  Add a DEA Number for the Doctor user for the eventual use during the
  writing of perscriptions.

  Add "SROSCH" key to Clerk users to allow them to schedule surgery
  appointments.

  Add "SROSCHOP" as secondary menu to all clerk users. This is the
  Surgery Scheduling Operations menu.

  Add "ONCO #SITE MANAGER MENU" to put a patient into Oncology suspense.

Add the necssary functions to the OSEHRASetup file and call the functions
from the PostImportSetupScript.

Change the setup of the users to not rely on hard-coded values in the
PostImportSetupScript, but to instead read from a CSV file.

Expand the list of users that are added by adding to the CSV file.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-157
Change-Id: Ied9e090a43b8d5252140558b283771d22d067e29