Update VistA infratructure for realism
Update the VistA setup based upon comments in the "Refactoring Test
Platforms" document.

Stop the HINQ call in the MAS PARAMETERS file.

Add a pharmacy OUTPATIENT SITE which uses the computer to assign Rx
numbers.  Also, set upper and lower bounds on the prescription numbers for
the outpatient site.

Add an Operating Room to the system so that surgical procedures can be
performed there.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-157
Change-Id: I1ca51f67071e3e2585659fd49b214e8cb52676ed