Expand set and reach of Patient list
Add to the number of patients that are added during the
PostImportSetupScript.  Be sure to reconnect to VistA between each
addition as the GT.M systems will error out while adding a patient.

Add extra functions to OSEHRASetup to be called
during the setup script to expand the patient entries to other package
specific patient files:

Oncology through the Casefinding/Suspense menu.  Ensure that you wait
for "Select" when waiting for the DHCP Menu to eliminate the possiblity
that the echo of the selection would trigger the wait function.

Surgery through the making of a surgical appointment.

Alter the ADTAction script to allow a greater number of lines to be
written on the "Log of Dispositions".  This eliminates the potential
pagination of the result log.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-157
Change-Id: Iacdb55c3b1039b02a78cd8b7f8588398a3bf3b9f