Expand set and reach of Patient list
Add to the number of patients that are added during the
PostImportSetupScript.  Add extra functions to OSEHRASetup to be called
during the setup script to expand the patient entries to other package
specific patient files:

Oncology through the Casefinding/Suspense menu
Surgery through the making of a surgical appointment.

Add all patients to be added to the system to the patdata0.csv file

Change the order of the import so that the patients will be added to the
oncology and surgery files during the postimportsetupscript.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-157
Change-Id: Iacdb55c3b1039b02a78cd8b7f8588398a3bf3b9f