Fix SplitZWR bug around globals with no file number
When the splitting happens, the script assumes that the file name will
have a FileMan File attached with a `+`.  If that isn't found, substitute
the number '0' and take the rest of the global name.

Update the PopulatePackages script to check the name of all globals for an
appropriate namespace. Attempt to split the name at the '+' sign and use
the appropriate value for the global name.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-193
Change-Id: I1e7e0532473078927b6dbc41e1d910fadd41dac3