Update ZGO to write new HEADER
Update the ZWR files that are written by the ZGO routine to
work with the latest version of GT.M.  This includes adding more
information to the second line.  The new information will take the date
and time of the writing of the .zwr file.

Change the "empty" global handling to:

OSEHRA-Id: http://issues.osehra.org/browse/OAT-216
Change-Id: Ie42466b1d0199cc8634695ea6a95853e9c854d23