Add a second connection to Autopatching
Carry a second connection through the auto-patching framework so that
simultaneous events can occur duing the install of a KIDS file.

Remove the passing of an unused list in the call within the
OTJKIDSUtilities.py file

This is necessary for the VA FileMan 22.2 install, as the Logons need
to be inhibited before it can be installed from the menu system.

Change-Id: I3f4b4476a234cc15ece411d6be6ac99677c1c6e2