Add Expanded DOX information
Add information about the Data Dictionaries, options, and functions to the
correct pages.  This information should be gathered from an XINDEX run on
the package itself and displayed as the other pieces of routine
information

Ensure that the lastest XINDEX patch has been applied before running the
tests.

Switch the DUZ back to 1 as the .5 user doesn't have the access to the
menus needed.

Change-Id: I9afeae3a4e8f1d6496d9f2c07576e4aae00c3d88