There is a bug in the TDateTimePicker.
Seems that outside of a US locale, you cannot set .MaxDate twice to
different values. Setting it a second time to 0 or INFINITY makes the
control tell me that my date is greater than the maximum of 1899. I
don't have time to take a deeper look at this, so for now, just make
sure to set the .MaxDate once.